# sandaga-backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sandaga App
  description: >
    'This is a ***Sandaga app API***.'
  termsOfService: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  contact:
    email: info@sandaga.com
  license:
    name: MIT
    url: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: https://sandaga.swagger.io/v2
...
paths:
...
  /api/v1/categories:
    get:
      tags:
        - categories
      summary: Returns all the matched categories
      description: Returns the categories that matches the given query criteria
      operationId: queryCategories
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags:
        - categories
      summary: Add a new category
      description: Creates a new category
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCategoryReq'
      responses:
        201:
          description: For successful user creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /api/v1/categories/{id}:
    get:
      tags:
        - categories
      summary: Returns a category
      description: Returns the category that matches the given category ID
      operationId: getCategory
      parameters:
        - name: id
          in: path
          description: Category Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      tags:
        - categories
      summary: Deletes category
      description: Deletes category based on given card ID.
      operationId: deleteCategoryById
      parameters:
        - name: id
          in: path
          description: category Identifier
          required: true
          schema:
            type: string
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: { }
    put:
      tags:
        - categories
      summary: Replace a category
      description: Replace a category with given id.
      operationId: replaceCategoryById
      parameters:
        - name: id
          in: path
          description: category Identifier
          required: true
          schema:
            type: string
      requestBody:
        description: Category object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        204:
          description: Category update successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Category'
        404:
          description: Given category ID doesn't exist
          content: { }
...
  /api/v1/images/{id}:
    delete:
      tags:
        - images
      summary: Deletes an image
      description: Deletes an image based on given image ID.
      operationId: deleteImageById
      parameters:
        - name: id
          in: path
          description: address Identifier
          required: true
          schema:
            type: string
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: { }
  /images/{fileName}:
    get:
      tags:
        - images
      summary: Returns an image
      description: Returns image based on given image fileName
      operationId: getImageByFileName
      parameters:
        - name: fileName
          in: path
          description: Image Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                  type: string
                  format: byte
  /api/v1/images:
    get:
      tags:
        - images
      summary: Returns all images
      description: Returns all images, else empty collection
      operationId: getAllImages
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
    post:
      tags:
        - images
      summary: Add a new image
      description: Creates a new image
      operationId: addImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageReq'
      responses:
        201:
          description: For successful user creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
  /api/v1/points:
    get:
      tags:
        - points
      summary: Returns all points
      description: Returns all points, else empty collection
      operationId: getAllPoints
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Point'
    post:
      tags:
        - points
      summary: Add a new point
      description: Creates a new point
      operationId: addPoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointReq'
      responses:
        201:
          description: For successful user creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Point'
  /api/v1/points/{id}:
    get:
      tags:
        - points
      summary: Returns an point
      description: Returns point based on given point ID
      operationId: getPointById
      parameters:
        - name: id
          in: path
          description: Point Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Point'
    put:
      tags:
        - points
      summary: Replace a Point
      description: Replace an point with given id.
      operationId: replacePointById
      parameters:
        - name: id
          in: path
          description: point Identifier
          required: true
          schema:
            type: string
      requestBody:
        description: Point object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Point'
      responses:
        204:
          description: Point update successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Point'
        404:
          description: Given point ID doesn't exist
          content: { }
    delete:
      tags:
        - points
      summary: Deletes an point
      description: Deletes an point based on given point ID.
      operationId: deletePointById
      parameters:
        - name: id
          in: path
          description: point Identifier
          required: true
          schema:
            type: string
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: { }
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        name:
          type: string
        channel:
          type: string
        subcategories:
            type: array
            items:
              $ref: '#/components/schemas/Subcategory'
      xml:
        name: Category
    AddCategoryReq:
      description: Request body for category
      type: object
      properties:
        label:
          description: label of the Category
          type: string
        name:
          description: name of the Category
          type: string
        channel:
          description: channel of the Category
          type: string
    Subcategory:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        name:
          type: string
        channel:
          type: string
      xml:
        name: Subcategory
...
    Ads:
      type: object
      properties:
        id:
          type: string
        ad_type:
          type: string
        subject:
          type: string
        type:
          type: string
        body:
          type: string
        donation:
          type: boolean
        price_cents:
          type: integer
          format: int32
        price_reco:
          type: integer
          format: int32
        shipping_type:
          type: string
        shipping_cost:
          type: string
        email:
          type: string
        phone:
          type: string
        phone_hidden_information_text:
          type: boolean
        item_condition:
          type: string
        video_game_type:
          type: string
        console_model:
          type: string
        console_brand:
          type: string
        image_sound_type_of_product:
          type: string
        image_sound_product:
          type: string
        phone_brand:
          type: string
        phone_color:
          type: string
        phone_model:
          type: string
        phone_memory:
          type: string
        furniture_type:
          type: string
        furniture_material:
          type: string
        furniture_color:
          type: string
        home_appliance_type:
          type: string
        home_appliance_product:
          type: string
        home_appliance_brand:
          type: string
        table_art_product:
          type: string
        table_art_material:
          type: string
        decoration_type:
          type: string
        linens_type:
          type: string
        linens_product:
          type: string
        linens_material:
          type: string
        diy_type:
          type: string
        diy_product:
          type: string
        gardering_type:
          type: string
        gardering_product:
          type: string
        clothing_type:
          type: string
        clothing_st:
          type: string
        clothing_category:
          type: string
        clothing_brand:
          type: string
        clothing_color:
          type: string
        clothing_condition:
          type: string
        shoe_type:
          type: string
        shoe_category:
          type: string
        shoe_size:
          type: string
        shoe_brand:
          type: string
        accessories_univers:
          type: string
        accessories_type:
          type: string
        accessories_brand:
          type: string
        accessories_material:
          type: string
        watches_jewels_brand:
          type: string
        watches_jewels_type:
          type: string
        watches_jewels_material:
          type: string
        baby_equipment_type:
          type: string
        baby_equipment_brand:
          type: string
        baby_clothing_category:
          type: string
        baby_clothing_brand:
          type: string
        baby_age:
          type: string
        bicycle_type:
          type: string
        bicycle_size:
          type: string
        bicycode:
          type: string
        sports_hobbies_activity:
          type: string
        toy_age:
          type: string
        toy_type:
          type: string
        matpro_agriculture_equipment_type:
          type: string
        regdate:
          type: string
      xml:
        name: Ads
...
    Address:
      type: object
      properties:
        id:
          description: Address identifier
          type: string
        number:
          description: house of flat number
          type: string
        street:
          description: street name
          type: string
        city:
          description: city name
          type: string
        state:
          description: state name
          type: string
        country:
          description: country name
          type: string
        pincode:
          description: postal code
          type: string
      xml:
        name: Address
...
    Image:
      type: object
      properties:
        id:
          description: Image identifier
          type: string
        name:
          type: string
        type:
          type: string
        imageUrl:
          description: Ads image URL
          type: string
        ads:
          $ref: '#/components/schemas/Ads'
      xml:
        name: Address
    ImageReq:
      type: object
      properties:
        file:
          type: string
          format: binary
        adsId:
          type: string
      xml:
        name: Image
...
    Point:
      type: object
      properties:
        id:
          description: Address identifier
          type: string
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        address:
          $ref: '#/components/schemas/Address'
      xml:
        name: Point
    PointReq:
      type: object
      properties:
        id:
          description: Address identifier
          type: string
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        addressId:
          type: string
      xml:
        name: Point
...